requirements.txt
# CrewAI + RAG (Ollama ile yerel çalışma)
crewai>=1.9.3
crewai-tools>=1.9.3

# Arayüz
streamlit>=1.28.0

# Ses → metin (transkript)
openai-whisper>=20231117
